+++ src/main/java/in/koreatech/koin/domain/order/shop/model/domain/ShopBaseDeliveryTips.java
@@ -8,7 +8,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import in.koreatech.koin.domain.order.shop.model.entity.ShopBaseDeliveryTip;
+import in.koreatech.koin.domain.order.shop.model.entity.delivery.ShopBaseDeliveryTip;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.OneToMany;
 import lombok.Getter;
