--- src/main/java/in/koreatech/koin/domain/order/shop/model/entity/ShopOperation.java
@@ -1,42 +1,0 @@
-package in.koreatech.koin.domain.order.shop.model.entity;
-
-import static jakarta.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
-
-import org.hibernate.annotations.Where;
-
-import in.koreatech.koin.domain.shop.model.shop.Shop;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = PROTECTED)
-@Table(name = "shop_operation")
-@Where(clause = "is_deleted=0")
-public class ShopOperation {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    @Column(name = "id", nullable = false, updatable = false)
-    private Integer id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "shop_id", nullable = false, unique = true)
-    private Shop shop;
-
-    @Column(name = "is_open", nullable = false)
-    private boolean isOpen = false;
-
-    @Column(name = "is_deleted", nullable = false)
-    private boolean isDeleted = false;
-}
