+++ src/main/java/in/koreatech/koin/domain/order/shop/model/entity/shop/ShopOperation.java
@@ -1,11 +1,10 @@
-package in.koreatech.koin.domain.order.shop.model.entity;
+package in.koreatech.koin.domain.order.shop.model.entity.shop;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 import org.hibernate.annotations.Where;
 
-import in.koreatech.koin._common.model.BaseEntity;
 import in.koreatech.koin.domain.shop.model.shop.Shop;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -14,6 +13,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -21,25 +21,22 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-@Table(name = "shop_base_delivery_tip")
+@Table(name = "shop_operation")
 @Where(clause = "is_deleted=0")
-public class ShopBaseDeliveryTip extends BaseEntity {
+public class ShopOperation {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id", nullable = false, updatable = false)
     private Integer id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "shop_id", nullable = false)
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "shop_id", nullable = false, unique = true)
     private Shop shop;
 
-    @Column(name = "order_amount", nullable = false)
-    private Integer orderAmount;
-
-    @Column(name = "fee", nullable = false)
-    private Integer fee;
+    @Column(name = "is_open", nullable = false)
+    private boolean isOpen = false;
 
     @Column(name = "is_deleted", nullable = false)
-    private Boolean isDeleted;
+    private boolean isDeleted = false;
 }
